requirements.txt
Flask
elasticsearch
google-generativeai
Flask-Cors
gunicorn
ijson